+++ robot-core/src/main/java/org/obolibrary/robot/Template.java
@@ -2184,9 +2184,14 @@ public class Template {
    * @return characteristics
    */
   private List<String> getCharacteristics(List<String> row) {
-    if (characteristicColumn != -1) {
-      String characteristicString = row.get(characteristicColumn);
-      if (characteristicSplit != null && characteristicString.contains(characteristicSplit)) {
+    if (characteristicColumn != -1 && characteristicColumn <= (row.size() - 1)) {
+      // If characteristics is the last column and there are no characteristics for
+      // this entry, the list will be one element short.
+      String characteristicString = row.get(characteristicColumn).trim();
+      if (characteristicString.equalsIgnoreCase("")) {
+        return new ArrayList<>();
+      } else if (characteristicSplit != null
+          && characteristicString.contains(characteristicSplit)) {
         return Arrays.asList(characteristicString.split(Pattern.quote(characteristicSplit)));
       } else {
         return Collections.singletonList(characteristicString.trim());