+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Now handles empty characteristics [#716]
+
+### Fixed
+- Handle empty [`template`] property charactersitics in [#717]
 
 ## [1.7.0] - 2020-07-31
 
@@ -192,6 +194,7 @@ First official release of ROBOT!
 [`report`]: http://robot.obolibrary.org/report
 [`template`]: http://robot.obolibrary.org/template
 
+[#717]: https://github.com/ontodev/robot/pull/716
 [#715]: https://github.com/ontodev/robot/pull/715
 [#710]: https://github.com/ontodev/robot/pull/710
 [#709]: https://github.com/ontodev/robot/issues/709